requirements.txt
# LangChain (stack 1.x; langchain-google-genai 4.x usa pacote google-genai)
langchain-core>=1.2.5,<2.0.0
langchain>=1.0.0
langchain-community>=0.4.0
langchain-openai>=0.3.0
langchain-text-splitters>=1.0.0

# LangSmith
langsmith>=0.3.45

# LLM Providers
langchain-google-genai>=4.0.0,<5.0.0
google-genai>=1.56.0,<2.0.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
pydantic>=2.0.0,<3.0.0

# Testing
pytest==8.3.4
